requirements.txt
#create env: conda create --name <env name> python=3.9
#install packages: pip install -r requirements.txt
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.5
cycler==0.11.0
dnspython==2.2.1
Flask==2.0.2
fonttools==4.28.5
gunicorn==20.1.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pymongo==4.3.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions==4.1.1
urllib3==1.26.13
Werkzeug==2.0.2
wincertstore==0.2